Запись при добавлении и редактировании

+++ module3/3/filework.h
@@ -4,10 +4,11 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <fcntl.h>
+#include <unistd.h>
 #include "contact.h"
 
-int read_contacts(const char *filename, contact **contacts, uint32_t *quan);
-long get_contacts_quan(int fd);
-int write_contact_to_file(int fd, const contact *c, uint32_t ind);
+int read_contacts(const char *filename, contact **contacts, uint64_t *quan);
+off_t get_contacts_quan(int fd);
+int write_contact_to_file(int fd, const contact *c, uint64_t ind);
 
 #endif
